// src/openapi/generated/models/paginated-response-resident-out.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ResidentOut } from '../models/resident-out';
export interface PaginatedResponseResidentOut {
  has_next: boolean;
  has_prev: boolean;
  items: Array<ResidentOut>;
  page: number;
  pages: number;
  size: number;
  total: number;
}